+++ scalr/data_source_workspace_ids_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/hashicorp/terraform/helper/resource"
 )
 
-func TestAccTFEWorkspaceIDsDataSource_basic(t *testing.T) {
+func TestAccScalrWorkspaceIDsDataSource_basic(t *testing.T) {
 	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
 
 	resource.Test(t, resource.TestCase{
@@ -17,7 +17,7 @@ func TestAccTFEWorkspaceIDsDataSource_basic(t *testing.T) {
 		Providers: testAccProviders,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccTFEWorkspaceIDsDataSourceConfig_basic(rInt),
+				Config: testAccScalrWorkspaceIDsDataSourceConfig_basic(rInt),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr(
 						"data.scalr_workspace_ids.foobar", "names.#", "2"),
@@ -26,25 +26,13 @@ func TestAccTFEWorkspaceIDsDataSource_basic(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"data.scalr_workspace_ids.foobar", "names.1", fmt.Sprintf("workspace-bar-%d", rInt)),
 					resource.TestCheckResourceAttr(
-						"data.scalr_workspace_ids.foobar", "organization", "existing-org"),
+						"data.scalr_workspace_ids.foobar", "environment_id", "existing-env"),
 					resource.TestCheckResourceAttr(
 						"data.scalr_workspace_ids.foobar", "ids.%", "2"),
-					resource.TestCheckResourceAttr(
-						"data.scalr_workspace_ids.foobar",
-						fmt.Sprintf("ids.workspace-foo-%d", rInt),
-						fmt.Sprintf("existing-org/workspace-foo-%d", rInt),
-					),
-					resource.TestCheckResourceAttr(
-						"data.scalr_workspace_ids.foobar",
-						fmt.Sprintf("ids.workspace-bar-%d", rInt),
-						fmt.Sprintf("existing-org/workspace-bar-%d", rInt),
-					),
-					resource.TestCheckResourceAttr(
-						"data.scalr_workspace_ids.foobar", "external_ids.%", "2"),
 					resource.TestCheckResourceAttrSet(
-						"data.scalr_workspace_ids.foobar", fmt.Sprintf("external_ids.workspace-foo-%d", rInt)),
+						"data.scalr_workspace_ids.foobar", fmt.Sprintf("ids.workspace-foo-%d", rInt)),
 					resource.TestCheckResourceAttrSet(
-						"data.scalr_workspace_ids.foobar", fmt.Sprintf("external_ids.workspace-bar-%d", rInt)),
+						"data.scalr_workspace_ids.foobar", fmt.Sprintf("ids.workspace-bar-%d", rInt)),
 					resource.TestCheckResourceAttrSet("data.scalr_workspace_ids.foobar", "id"),
 				),
 			},
@@ -52,7 +40,7 @@ func TestAccTFEWorkspaceIDsDataSource_basic(t *testing.T) {
 	})
 }
 
-func TestAccTFEWorkspaceIDsDataSource_wildcard(t *testing.T) {
+func TestAccScalrWorkspaceIDsDataSource_wildcard(t *testing.T) {
 	t.Skip("Wildcard test is not passing for unknown reasons. Using the wildcard symbol produces no workspaces")
 	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
 
@@ -61,39 +49,22 @@ func TestAccTFEWorkspaceIDsDataSource_wildcard(t *testing.T) {
 		Providers: testAccProviders,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccTFEWorkspaceIDsDataSourceConfig_wildcard(rInt),
+				Config: testAccScalrWorkspaceIDsDataSourceConfig_wildcard(rInt),
 				Check: resource.ComposeAggregateTestCheckFunc(
 					resource.TestCheckResourceAttr(
 						"data.scalr_workspace_ids.foobar", "names.#", "1"),
 					resource.TestCheckResourceAttr(
 						"data.scalr_workspace_ids.foobar", "names.0", "*"),
 					resource.TestCheckResourceAttr(
-						"data.scalr_workspace_ids.foobar", "organization", "existing-org"),
+						"data.scalr_workspace_ids.foobar", "environment_id", "existing-env"),
 					resource.TestCheckResourceAttr(
 						"data.scalr_workspace_ids.foobar", "ids.%", "3"),
-					resource.TestCheckResourceAttr(
-						"data.scalr_workspace_ids.foobar",
-						fmt.Sprintf("ids.workspace-foo-%d", rInt),
-						fmt.Sprintf("existing-org/workspace-foo-%d", rInt),
-					),
-					resource.TestCheckResourceAttr(
-						"data.scalr_workspace_ids.foobar",
-						fmt.Sprintf("ids.workspace-bar-%d", rInt),
-						fmt.Sprintf("existing-org/workspace-bar-%d", rInt),
-					),
-					resource.TestCheckResourceAttr(
-						"data.scalr_workspace_ids.foobar",
-						fmt.Sprintf("ids.workspace-dummy-%d", rInt),
-						fmt.Sprintf("existing-org/workspace-dummy-%d", rInt),
-					),
-					resource.TestCheckResourceAttr(
-						"data.scalr_workspace_ids.foobar", "external_ids.%", "3"),
 					resource.TestCheckResourceAttrSet(
-						"data.scalr_workspace_ids.foobar", fmt.Sprintf("external_ids.workspace-foo-%d", rInt)),
+						"data.scalr_workspace_ids.foobar", fmt.Sprintf("ids.workspace-foo-%d", rInt)),
 					resource.TestCheckResourceAttrSet(
-						"data.scalr_workspace_ids.foobar", fmt.Sprintf("external_ids.workspace-bar-%d", rInt)),
+						"data.scalr_workspace_ids.foobar", fmt.Sprintf("ids.workspace-bar-%d", rInt)),
 					resource.TestCheckResourceAttrSet(
-						"data.scalr_workspace_ids.foobar", fmt.Sprintf("external_ids.workspace-dummy-%d", rInt)),
+						"data.scalr_workspace_ids.foobar", fmt.Sprintf("ids.workspace-dummy-%d", rInt)),
 					resource.TestCheckResourceAttrSet("data.scalr_workspace_ids.foobar", "id"),
 				),
 			},
@@ -101,48 +72,48 @@ func TestAccTFEWorkspaceIDsDataSource_wildcard(t *testing.T) {
 	})
 }
 
-func testAccTFEWorkspaceIDsDataSourceConfig_basic(rInt int) string {
+func testAccScalrWorkspaceIDsDataSourceConfig_basic(rInt int) string {
 	return fmt.Sprintf(`
 resource "scalr_workspace" "foo" {
-  name         = "workspace-foo-%d"
-  organization = "existing-org"
+  name           = "workspace-foo-%d"
+  environment_id = "existing-env"
 }
 
 resource "scalr_workspace" "bar" {
-  name         = "workspace-bar-%d"
-  organization = "existing-org"
+  name           = "workspace-bar-%d"
+  environment_id = "existing-env"
 }
 
 resource "scalr_workspace" "dummy" {
-  name         = "workspace-dummy-%d"
-  organization = "existing-org"
+  name           = "workspace-dummy-%d"
+  environment_id = "existing-env"
 }
 
 data "scalr_workspace_ids" "foobar" {
-  names        = ["${scalr_workspace.foo.name}", "${scalr_workspace.bar.name}"]
-  organization = "existing-org"
+  names          = ["${scalr_workspace.foo.name}", "${scalr_workspace.bar.name}"]
+  environment_id = "existing-env"
 }`, rInt, rInt, rInt)
 }
 
-func testAccTFEWorkspaceIDsDataSourceConfig_wildcard(rInt int) string {
+func testAccScalrWorkspaceIDsDataSourceConfig_wildcard(rInt int) string {
 	return fmt.Sprintf(`
 resource "scalr_workspace" "foo" {
-  name         = "workspace-foo-%d"
-  organization = "existing-org"
+  name           = "workspace-foo-%d"
+  environment_id = "existing-env"
 }
 
 resource "scalr_workspace" "bar" {
-  name         = "workspace-bar-%d"
-  organization = "existing-org"
+  name           = "workspace-bar-%d"
+  environment_id = "existing-env"
 }
 
 resource "scalr_workspace" "dummy" {
-  name         = "workspace-dummy-%d"
-  organization = "existing-org"
+  name           = "workspace-dummy-%d"
+  environment_id = "existing-env"
 }
 
 data "scalr_workspace_ids" "foobar" {
-  names        = ["*"]
-  organization = "existing-org"
+  names          = ["*"]
+  environment_id = "existing-env"
 }`, rInt, rInt, rInt)
 }