+++ scalr/resource_scalr_workspace_test.go
@@ -11,100 +11,20 @@ import (
 	scalr "github.com/scalr/go-scalr"
 )
 
-func TestPackWorkspaceID(t *testing.T) {
-	cases := []struct {
-		w   *scalr.Workspace
-		id  string
-		err bool
-	}{
-		{
-			w: &scalr.Workspace{
-				Name: "my-workspace-name",
-				Organization: &scalr.Organization{
-					Name: "my-org-name",
-				},
-			},
-			id:  "my-org-name/my-workspace-name",
-			err: false,
-		},
-		{
-			w: &scalr.Workspace{
-				Name: "my-workspace-name",
-			},
-			id:  "",
-			err: true,
-		},
-	}
-
-	for _, tc := range cases {
-		id, err := packWorkspaceID(tc.w)
-		if (err != nil) != tc.err {
-			t.Fatalf("expected error is %t, got %v", tc.err, err)
-		}
-
-		if tc.id != id {
-			t.Fatalf("expected ID %q, got %q", tc.id, id)
-		}
-	}
-}
-
-func TestUnpackWorkspaceID(t *testing.T) {
-	cases := []struct {
-		id   string
-		org  string
-		name string
-		err  bool
-	}{
-		{
-			id:   "my-org-name/my-workspace-name",
-			org:  "my-org-name",
-			name: "my-workspace-name",
-			err:  false,
-		},
-		{
-			id:   "my-workspace-name|my-org-name",
-			org:  "my-org-name",
-			name: "my-workspace-name",
-			err:  false,
-		},
-		{
-			id:   "some-invalid-id",
-			org:  "",
-			name: "",
-			err:  true,
-		},
-	}
-
-	for _, tc := range cases {
-		org, name, err := unpackWorkspaceID(tc.id)
-		if (err != nil) != tc.err {
-			t.Fatalf("expected error is %t, got %v", tc.err, err)
-		}
-
-		if tc.org != org {
-			t.Fatalf("expected organization %q, got %q", tc.org, org)
-		}
-
-		if tc.name != name {
-			t.Fatalf("expected name %q, got %q", tc.name, name)
-		}
-	}
-}
-
-func TestAccTFEWorkspace_basic(t *testing.T) {
+func TestAccScalrWorkspace_basic(t *testing.T) {
 	workspace := &scalr.Workspace{}
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
 		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
+		CheckDestroy: testAccCheckScalrWorkspaceDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccTFEWorkspace_basic,
+				Config: testAccScalrWorkspace_basic,
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckTFEWorkspaceExists(
+					testAccCheckScalrWorkspaceExists(
 						"scalr_workspace.foobar", workspace),
-					testAccCheckTFEWorkspaceAttributes(workspace),
+					testAccCheckScalrWorkspaceAttributes(workspace),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.foobar", "name", "workspace-test"),
 					resource.TestCheckResourceAttr(
@@ -115,7 +35,6 @@ func TestAccTFEWorkspace_basic(t *testing.T) {
 						"scalr_workspace.foobar", "queue_all_runs", "true"),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.foobar", "working_directory", ""),
-					resource.TestCheckResourceAttrSet("scalr_workspace.foobar", "external_id"),
 					resource.TestCheckResourceAttrSet("scalr_workspace.foobar", "created_by.0.full_name"),
 					resource.TestCheckResourceAttrSet("scalr_workspace.foobar", "created_by.0.email"),
 					resource.TestCheckResourceAttrSet("scalr_workspace.foobar", "created_by.0.username"),
@@ -125,20 +44,20 @@ func TestAccTFEWorkspace_basic(t *testing.T) {
 	})
 }
 
-func TestAccTFEWorkspace_monorepo(t *testing.T) {
+func TestAccScalrWorkspace_monorepo(t *testing.T) {
 	workspace := &scalr.Workspace{}
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
 		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
+		CheckDestroy: testAccCheckScalrWorkspaceDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccTFEWorkspace_monorepo,
+				Config: testAccScalrWorkspace_monorepo,
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckTFEWorkspaceExists(
+					testAccCheckScalrWorkspaceExists(
 						"scalr_workspace.foobar", workspace),
-					testAccCheckTFEWorkspaceMonorepoAttributes(workspace),
+					testAccCheckScalrWorkspaceMonorepoAttributes(workspace),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.foobar", "name", "workspace-monorepo"),
 					resource.TestCheckResourceAttr(
@@ -151,20 +70,20 @@ func TestAccTFEWorkspace_monorepo(t *testing.T) {
 	})
 }
 
-func TestAccTFEWorkspace_renamed(t *testing.T) {
+func TestAccScalrWorkspace_renamed(t *testing.T) {
 	workspace := &scalr.Workspace{}
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
 		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
+		CheckDestroy: testAccCheckScalrWorkspaceDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccTFEWorkspace_basic,
+				Config: testAccScalrWorkspace_basic,
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckTFEWorkspaceExists(
+					testAccCheckScalrWorkspaceExists(
 						"scalr_workspace.foobar", workspace),
-					testAccCheckTFEWorkspaceAttributes(workspace),
+					testAccCheckScalrWorkspaceAttributes(workspace),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.foobar", "name", "workspace-test"),
 					resource.TestCheckResourceAttr(
@@ -179,13 +98,13 @@ func TestAccTFEWorkspace_renamed(t *testing.T) {
 			},
 
 			{
-				PreConfig: testAccCheckTFEWorkspaceRename,
-				Config:    testAccTFEWorkspace_renamed,
+				PreConfig: testAccCheckScalrWorkspaceRename,
+				Config:    testAccScalrWorkspace_renamed,
 				PlanOnly:  true,
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckTFEWorkspaceExists(
+					testAccCheckScalrWorkspaceExists(
 						"scalr_workspace.foobar", workspace),
-					testAccCheckTFEWorkspaceAttributes(workspace),
+					testAccCheckScalrWorkspaceAttributes(workspace),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.foobar", "name", "workspace-test"),
 					resource.TestCheckResourceAttr(
@@ -201,20 +120,20 @@ func TestAccTFEWorkspace_renamed(t *testing.T) {
 		},
 	})
 }
-func TestAccTFEWorkspace_update(t *testing.T) {
+func TestAccScalrWorkspace_update(t *testing.T) {
 	workspace := &scalr.Workspace{}
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
 		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
+		CheckDestroy: testAccCheckScalrWorkspaceDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccTFEWorkspace_basic,
+				Config: testAccScalrWorkspace_basic,
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckTFEWorkspaceExists(
+					testAccCheckScalrWorkspaceExists(
 						"scalr_workspace.foobar", workspace),
-					testAccCheckTFEWorkspaceAttributes(workspace),
+					testAccCheckScalrWorkspaceAttributes(workspace),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.foobar", "name", "workspace-test"),
 					resource.TestCheckResourceAttr(
@@ -229,11 +148,11 @@ func TestAccTFEWorkspace_update(t *testing.T) {
 			},
 
 			{
-				Config: testAccTFEWorkspace_update,
+				Config: testAccScalrWorkspace_update,
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckTFEWorkspaceExists(
+					testAccCheckScalrWorkspaceExists(
 						"scalr_workspace.foobar", workspace),
-					testAccCheckTFEWorkspaceAttributesUpdated(workspace),
+					testAccCheckScalrWorkspaceAttributesUpdated(workspace),
 					resource.TestCheckResourceAttr(
 						"scalr_workspace.foobar", "name", "workspace-updated"),
 					resource.TestCheckResourceAttr(
@@ -252,14 +171,14 @@ func TestAccTFEWorkspace_update(t *testing.T) {
 	})
 }
 
-func TestAccTFEWorkspace_import(t *testing.T) {
+func TestAccScalrWorkspace_import(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
 		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
+		CheckDestroy: testAccCheckScalrWorkspaceDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccTFEWorkspace_basic,
+				Config: testAccScalrWorkspace_basic,
 			},
 
 			{
@@ -271,7 +190,7 @@ func TestAccTFEWorkspace_import(t *testing.T) {
 	})
 }
 
-func testAccCheckTFEWorkspaceExists(
+func testAccCheckScalrWorkspaceExists(
 	n string, workspace *scalr.Workspace) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		scalrClient := testAccProvider.Meta().(*scalr.Client)
@@ -285,24 +204,10 @@ func testAccCheckTFEWorkspaceExists(
 			return fmt.Errorf("No instance ID is set")
 		}
 
-		// Get the organization and workspace name.
-		organization, name, err := unpackWorkspaceID(rs.Primary.ID)
-		if err != nil {
-			return fmt.Errorf("Error unpacking workspace ID: %v", err)
-		}
-
-		w, err := scalrClient.Workspaces.Read(ctx, organization, name)
+		// Get the workspace
+		w, err := scalrClient.Workspaces.ReadByID(ctx, rs.Primary.ID)
 		if err != nil {
 			return err
-		}
-
-		id, err := packWorkspaceID(w)
-		if err != nil {
-			return fmt.Errorf("Error creating ID for workspace %s: %v", name, err)
-		}
-
-		if id != rs.Primary.ID {
-			return fmt.Errorf("Workspace not found")
 		}
 
 		*workspace = *w
@@ -311,7 +216,7 @@ func testAccCheckTFEWorkspaceExists(
 	}
 }
 
-func testAccCheckTFEWorkspaceAttributes(
+func testAccCheckScalrWorkspaceAttributes(
 	workspace *scalr.Workspace) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		if workspace.Name != "workspace-test" {
@@ -338,7 +243,7 @@ func testAccCheckTFEWorkspaceAttributes(
 	}
 }
 
-func testAccCheckTFEWorkspaceMonorepoAttributes(
+func testAccCheckScalrWorkspaceMonorepoAttributes(
 	workspace *scalr.Workspace) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		if workspace.Name != "workspace-monorepo" {
@@ -353,12 +258,12 @@ func testAccCheckTFEWorkspaceMonorepoAttributes(
 	}
 }
 
-func testAccCheckTFEWorkspaceRename() {
+func testAccCheckScalrWorkspaceRename() {
 	scalrClient := testAccProvider.Meta().(*scalr.Client)
 
 	w, err := scalrClient.Workspaces.Update(
 		context.Background(),
-		"existing-org",
+		"existing-env",
 		"workspace-test",
 		scalr.WorkspaceUpdateOptions{Name: scalr.String("renamed-out-of-band")},
 	)
@@ -371,7 +276,7 @@ func testAccCheckTFEWorkspaceRename() {
 	}
 }
 
-func testAccCheckTFEWorkspaceAttributesUpdated(
+func testAccCheckScalrWorkspaceAttributesUpdated(
 	workspace *scalr.Workspace) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		if workspace.Name != "workspace-updated" {
@@ -402,7 +307,7 @@ func testAccCheckTFEWorkspaceAttributesUpdated(
 	}
 }
 
-func testAccCheckTFEWorkspaceDestroy(s *terraform.State) error {
+func testAccCheckScalrWorkspaceDestroy(s *terraform.State) error {
 	scalrClient := testAccProvider.Meta().(*scalr.Client)
 
 	for _, rs := range s.RootModule().Resources {
@@ -414,13 +319,7 @@ func testAccCheckTFEWorkspaceDestroy(s *terraform.State) error {
 			return fmt.Errorf("No instance ID is set")
 		}
 
-		// Get the organization and workspace name.
-		organization, name, err := unpackWorkspaceID(rs.Primary.ID)
-		if err != nil {
-			return fmt.Errorf("Error unpacking workspace ID: %v", err)
-		}
-
-		_, err = scalrClient.Workspaces.Read(ctx, organization, name)
+		_, err := scalrClient.Workspaces.ReadByID(ctx, rs.Primary.ID)
 		if err == nil {
 			return fmt.Errorf("Workspace %s still exists", rs.Primary.ID)
 		}
@@ -429,31 +328,31 @@ func testAccCheckTFEWorkspaceDestroy(s *terraform.State) error {
 	return nil
 }
 
-const testAccTFEWorkspace_basic = `
+const testAccScalrWorkspace_basic = `
 resource "scalr_workspace" "foobar" {
-  name         = "workspace-test"
-  organization = "existing-org"
-  auto_apply   = true
+  name           = "workspace-test"
+  environment_id = "existing-env"
+  auto_apply     = true
 }`
 
-const testAccTFEWorkspace_monorepo = `
+const testAccScalrWorkspace_monorepo = `
 resource "scalr_workspace" "foobar" {
   name                  = "workspace-monorepo"
-  organization          = "existing-org"
+  environment_id        = "existing-env"
   working_directory     = "/db"
 }`
 
-const testAccTFEWorkspace_renamed = `
+const testAccScalrWorkspace_renamed = `
 resource "scalr_workspace" "foobar" {
-  name         = "renamed-out-of-band"
-  organization = "existing-org"
-  auto_apply   = true
+  name           = "renamed-out-of-band"
+  environment_id = "existing-env"
+  auto_apply     = true
 }`
 
-const testAccTFEWorkspace_update = `
+const testAccScalrWorkspace_update = `
 resource "scalr_workspace" "foobar" {
   name                  = "workspace-updated"
-  organization          = "existing-org"
+  environment_id        = "existing-env"
   auto_apply            = false
   operations            = false
   queue_all_runs        = false